Blocks now correctly fall when a block below it is removed
-Added global variables for screen width and height in confg.py
-Added a condition to the if statement that checks if a block should
fall in the wait method of block.py
-Used the global screen width in blockgrid.py
-You will see a lot of changes in blockgrid.py. These aren't actually
changes to the code, they are just formatting changes that I made from
updating my code to the latest commit. This isn't intentional, but I
didn't feel it was a huge issue to fix.
-One thing I noticed is sometimes a falling block will appear at it's
final position when a block below it is removed before it reaches the
final position. It doesn't always happen, so I'm not sure of the cause,
but maybe we can work on it when we meet. I can also show the problem
because a text explanation isn't that great.