Added a debug command and fixed highscores
-Added the ability to end the game because it makes it easier to test
adding various scores to the highscore table. I was originally going to
put this in the highscorestate but the class doesn't allow for this to
easily be implemented.
-The highscorestate had some errors.
-The normal mode table was passing in the wrong mode value.
-I had to change kwargs to args after doing some investigating. Args
contains a tuple made up of score and mode. This tuple is accessed via
args[0] then to access score you have to add ['score'] resulting in
args[0]['score']. I don't believe this is intended, but it works until
it is corrected.
-I am committing from my laptop so hopefully I changed all the correct
settings, so that the whole thing isn't committed. ingame.py: Line 86
and 107 highscore.py: line 27,77,140